models/vertex_ai_maxims/requirements.txt
dify_plugin>=0.2.0,<0.4.0
anthropic==0.42.0
google-cloud-aiplatform==1.97.0
google-genai==1.30.0
Pillow==11.0.0
